# docs/openapi.yaml
components:
    schemas:
...
        budget.GetBudgetUseCaseBudgetCategories:
            properties:
                amount:
                    type: integer
                available:
                    type: integer
                budget_id:
                    type: string
                category:
                    $ref: '#/components/schemas/entity.TransactionCategory'
                category_id:
                    type: string
                created_at:
                    type: string
                deleted_at:
                    type: string
                id:
                    type: string
                spent:
                    type: integer
                updated_at:
                    type: string
            type: object
        budget.UpsertBudgetUseCaseCategoryInput:
            properties:
                amount:
                    type: integer
                category_id:
                    type: string
            required:
                - amount
                - category_id
            type: object
...
        dateutil.ComparisonDates:
            properties:
                comparison_end_date:
                    type: string
                comparison_start_date:
                    type: string
                end_date:
                    type: string
                start_date:
                    type: string
            type: object
...
        dto.ErrorItem:
            properties:
                name:
                    type: string
                reason:
                    type: string
            type: object
        dto.ErrorResponse:
            properties:
                errors:
                    items:
                        $ref: '#/components/schemas/dto.ErrorItem'
                    type: array
                message:
                    type: string
            type: object
...
        dto.GetBudgetCategoryResponse:
            properties:
                amount:
                    type: integer
                available:
                    type: integer
                created_at:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                name:
                    type: string
                spent:
                    type: integer
                updated_at:
                    type: string
            type: object
        dto.GetBudgetResponse:
            properties:
                amount:
                    type: integer
                available:
                    type: integer
                available_per_day:
                    type: integer
                available_per_day_percentage_variation:
                    type: integer
                available_percentage_variation:
                    type: integer
                budget_categories:
                    items:
                        $ref: '#/components/schemas/budget.GetBudgetUseCaseBudgetCategories'
                    type: array
                comparison_dates:
                    $ref: '#/components/schemas/dateutil.ComparisonDates'
                created_at:
                    type: string
                date:
                    type: string
                deleted_at:
                    type: string
                id:
                    type: string
                spent:
                    type: integer
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
...
        dto.ListTransactionsResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/entity.FullTransaction'
                    type: array
                page:
                    type: integer
                page_size:
                    type: integer
                total_items:
                    type: integer
                total_pages:
                    type: integer
            type: object
...
        dto.UpsertBudgetRequest:
            properties:
                amount:
                    type: integer
                categories:
                    items:
                        $ref: '#/components/schemas/budget.UpsertBudgetUseCaseCategoryInput'
                    type: array
                date:
                    type: string
            required:
                - amount
                - date
            type: object
...
        entity.FullTransaction:
            properties:
                account_id:
                    type: string
                amount:
                    type: integer
                category_id:
                    type: string
                category_name:
                    type: string
                created_at:
                    type: string
                date:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                institution_id:
                    type: string
                institution_logo:
                    type: string
                institution_name:
                    type: string
                is_ignored:
                    type: boolean
                name:
                    type: string
                payment_method_id:
                    type: string
                payment_method_name:
                    type: string
                search_document: {}
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
...
        entity.TransactionCategory:
            properties:
                created_at:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                name:
                    type: string
                updated_at:
                    type: string
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            description: Type "Bearer" followed by a space and JWT token.
            in: header
            name: Authorization
            type: apiKey
info:
    contact:
        email: danielmesquitta123@gmail.com
        name: Daniel Mesquita
    description: API Finance Manager
    title: API Finance Manager
    version: "1.0"
openapi: 3.0.3
paths:
...
    /v1/budgets:
        delete:
            description: Delete budget
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Delete budget
            tags:
                - Budget
        get:
            description: Get budget
            parameters:
                - description: Date
                  in: query
                  name: date
                  required: true
                  schema:
                    format: date
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.GetBudgetResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get budget
            tags:
                - Budget
        post:
            description: Create or update budget
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.UpsertBudgetRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Create or update budget
            tags:
                - Budget
    /v1/budgets/categories/{category_id}:
        get:
            description: Get budget category
            parameters:
                - description: Category ID
                  in: path
                  name: category_id
                  required: true
                  schema:
                    format: uuid
                    type: string
                - description: Date
                  in: query
                  name: date
                  required: true
                  schema:
                    format: date
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.GetBudgetCategoryResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get budget category
            tags:
                - Budget
    /v1/budgets/categories/{category_id}/transactions:
        get:
            description: List budget category transactions
            parameters:
                - description: Category ID
                  in: path
                  name: category_id
                  required: true
                  schema:
                    format: uuid
                    type: string
                - description: Page
                  in: query
                  name: page
                  schema:
                    type: integer
                - description: Page size
                  in: query
                  name: page_size
                  schema:
                    type: integer
                - description: Date
                  in: query
                  name: date
                  required: true
                  schema:
                    format: date
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ListTransactionsResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List budget category transactions
            tags:
                - Budget
...